+++ CHANGELOG.md
@@ -8,7 +8,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 
 ### Added
 
-* Allow the component to be more customisable (#6)
+* Allow the component to be more customisable (#6, #8)
 * Add documentation around testing (#7)
 
 ## [0.2.0] - 2023-07-17